퀴즈 이미지를 사용중인 경우 퀴즈를 삭제하지 못하도록 처리 (#151)
* feat: 퀴즈 이미지를 사용중인 경우 퀴즈를 삭제하지 못하도록 처리

+++ src/modules/quiz-image/use-cases/delete-quiz-image/delete-quiz-image.handler.ts
@@ -3,12 +3,17 @@ import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 
 import { Transactional } from '@nestjs-cls/transactional';
 
+import { QuizImageInUsedError } from '@module/quiz-image/errors/quiz-image-in-used.error';
 import { QuizImageNotFoundError } from '@module/quiz-image/errors/quiz-image-not-found.error';
 import {
   QUIZ_IMAGE_REPOSITORY,
   QuizImageRepositoryPort,
 } from '@module/quiz-image/repositories/quiz-image/quiz-image.repository.port';
 import { DeleteQuizImageCommand } from '@module/quiz-image/use-cases/delete-quiz-image/delete-quiz-image.command';
+import {
+  QUIZ_REPOSITORY,
+  QuizRepositoryPort,
+} from '@module/quiz/repositories/quiz/quiz.repository.port';
 
 import { AWS_S3_PORT, AwsS3Port } from '@shared/services/aws-s3/aws-s3.port';
 
@@ -24,6 +29,8 @@ export class DeleteQuizImageHandler
   constructor(
     @Inject(QUIZ_IMAGE_REPOSITORY)
     private readonly quizImageRepository: QuizImageRepositoryPort,
+    @Inject(QUIZ_REPOSITORY)
+    private readonly quizRepository: QuizRepositoryPort,
     @Inject(AWS_S3_PORT)
     private readonly awsS3Adapter: AwsS3Port,
     @Inject(EVENT_STORE)
@@ -38,6 +45,14 @@ export class DeleteQuizImageHandler
 
     if (existingQuizImage === undefined) {
       throw new QuizImageNotFoundError();
+    }
+
+    const quizzesUsingImage = await this.quizRepository.findManyByFileNames(
+      new Set([existingQuizImage.fileName]),
+    );
+
+    if (quizzesUsingImage.length > 0) {
+      throw new QuizImageInUsedError();
     }
 
     existingQuizImage.delete();
